Resolver los tres avisos restantes de React Doctor
Los tres eran legítimos, aunque cosméticos comparados con los bugs anteriores:

- `handleRemoveBadge` (P2P.jsx) no leía nada del componente, así que se
  reconstruía en cada render sin motivo y rompía la memoización de los hijos
  que la reciben por prop. Sube a scope de módulo con su comentario.

- `formatCoinAmount` y `formatCoinPrice` vivían en QPCoinRow.jsx. Un archivo
  que exporta cosas que no son componentes impide a Fast Refresh preservar el
  estado: cada edición del archivo forzaba una recarga completa del módulo.
  Se mueven a `helpers/coinFormat.js` y QPCoinRow los importa.

Los tests de formato viajan con las funciones. Sin cambios de comportamiento.
